+++ src/ValueObject/Route/V2/Product.php
@@ -19,7 +19,9 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
 
-namespace Shopgate\CloudIntegrationSdk\ValueObject\Route;
+namespace Shopgate\CloudIntegrationSdk\ValueObject\Route\V2;
+
+use Shopgate\CloudIntegrationSdk\ValueObject\Route\AbstractRoute;
 
 class Product extends AbstractRoute
 {
@@ -27,6 +29,6 @@ class Product extends AbstractRoute
     {
         $this->paramNameList = array('productId');
         $this->identifier    = "/products/{$this->paramNameList[0]}";
-        $this->pattern       = "~\/products\/(?<{$this->paramNameList[0]}>[^/?]*)\/*$~";
+        $this->pattern       = "~^\/v2\/products\/(?<{$this->paramNameList[0]}>[^/?]*)\/*$~";
     }
 }
